more shit. almost kinda works now.
kinda got the logic going that tries to fit plays on to the
board, checks for bogus words after, and loops.
that has lots of fussy breaking stuff up in to different
strings.
it was making some good plays but then it's failing to stuff
and not finding any play when many obvious ones are available
and the word should be matching the line pattern.

subpermutations returns the number of results in the 0th element.
subpermutations2 does elaborate shit to try to check that a word will fit
but has some kind of a problem with the pattern passed in not matching when
it should.
new move_tile_to_board() translates from simple x,y to pixel coordinates.
dump_board() takes a board param but it's kinda redundant now.
build_board_rows_and_cols_arrays() -> build_board_lines_arrays().
instead of building rows and cols in separate arrays, just put them in
the same ones.  iterating over only one array simplifies other stuff.
also the board_lines_pat it creates matches '*' as a single-char wildcard.
that was needed to actually get our plays to match in subpermutations2.
otherwise nothing was matching as the strings weren't long enough (or if
they were, they were wrong).
count_invalid_words() takes the build_lines_text as generated by
build_board_rows_and_cols_arrays() as a param now and properly localizes
join_array() gets an optional separater so that eg build_lines_text
can be nicely output with one line per line.
play_on_board() renders a play to the board, which is passed in, and
optionally sends the tile moves to the server as well.