scab-client.sno:127: Error 4 in statement 165 at level 2 Null string in illegal context
that was caused by trying to access tile.letter(null), which was
caused by a few bugs in draw_rack().
draw_rack() added one to the size of the list of sushi (tiles
available to be drawn, named after the pattern on the little cloth
purse we used to hold the tiles in real life... "please pass the
sushi"... "are there any sushi left?") rather than the modulo-
random number which needed 1 added to it to make it 1-indexed.
so that was letting modulo-randoms try to access the 0th
element which doesn't exist, and moving the end boundary down
further, creating (especially near the end of game when sushi
are low) a chance that it would try to pick a tile from an empty
slot, then on top of all of that, when it was picking a tile,
it wasn't removing that item.  so shuffling the last item down
to replace the removed item should hopefully keep that list
of sushi consistent.